+++ internal/kml/kmltrack.go
@@ -52,7 +52,11 @@ func (gxt *TrackBuilderEnsemble) Generate(aeroTrack *aeroapi.Track) (*Track, err
 
 	kmlAssets := make(map[string]any)
 	for _, kb := range gxt.Builders {
-		kmlThing := kb.Build(positions)
+		kmlThing, buildErr := kb.Build(positions)
+		if buildErr != nil {
+			fmt.Printf("NOTE: %s\n", buildErr)
+			continue
+		}
 		mainDocument.Append(kmlThing.Root)
 		for k, v := range kmlThing.Assets {
 			kmlAssets[k] = v