+++ internal/os/launchfile.go
@@ -2,8 +2,17 @@ package os
 
 import (
     "os/exec"
+    "runtime"
 )
 
 func LaunchFile(filename string) error {
-    return exec.Command("open", filename).Run()
+    switch runtime.GOOS {
+    case "windows":
+        return exec.Command("cmd", "/C", filename).Run()
+    case "darwin":
+        return exec.Command("open", filename).Run()
+    case "linux":
+        return exec.Command("sh", "-c", filename).Run()
+    }
+    return exec.Command(filename).Run()
 }