+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
-## [pre-v0.0.1](https://github.com/noodnik2/flightvisualizer/compare/main...mock-output-folder) (2023-05-24)
-* pre-release modifications
+## [v0.0.1](https://github.com/noodnik2/flightvisualizer/releases/tag/v0.0.1) (2023-06-02)
+* Initial Release
+
+## v0.0.2 (2023-06-05)
+* Addressed some `TODO` items
+* Added top-level `--version` and `--debug` flags
+* Fixed launch & set environment bugs in Windows